backend/requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
google-generativeai==0.3.1
google-adk==0.0.3
Flask-Cors==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
